refactor(source): split parser into separate crate (part 1)
