+++ src/connector/encdec/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
-name = "risingwave_connector"
+name = "risingwave_connector_encdec"
+description = "Encoding and decoding between external data formats and RisingWave datum"
 version = { workspace = true }
 edition = { workspace = true }
 homepage = { workspace = true }
@@ -112,7 +113,6 @@ regex = "1.4"
 reqwest = { version = "0.12.2", features = ["json", "stream"] }
 risingwave_common = { workspace = true }
 risingwave_common_estimate_size = { workspace = true }
-risingwave_connector_encdec = { workspace = true }
 risingwave_jni_core = { workspace = true }
 risingwave_object_store = { workspace = true }
 risingwave_pb = { workspace = true }
@@ -152,11 +152,11 @@ tracing = "0.1"
 url = "2"
 urlencoding = "2"
 uuid = { version = "1", features = ["v4", "fast-rng"] }
-with_options = { path = "./with_options" }
+with_options = { path = "../with_options" }
 yup-oauth2 = "8.3"
 
 [target.'cfg(not(madsim))'.dependencies]
-workspace-hack = { path = "../workspace-hack" }
+workspace-hack = { path = "../../workspace-hack" }
 
 [dev-dependencies]
 assert_matches = "1"
@@ -178,22 +178,6 @@ walkdir = "2"
 [build-dependencies]
 prost-build = "0.12"
 protobuf-src = "1"
-
-[[bench]]
-name = "debezium_json_parser"
-harness = false
-
-[[bench]]
-name = "nexmark_integration"
-harness = false
-
-[[bench]]
-name = "json_parser_case_insensitive"
-harness = false
-
-[[bench]]
-name = "json_vs_plain_parser"
-harness = false
 
 [lints]
 workspace = true
