+++ src/connector/src/parser/json_parser.rs
@@ -102,7 +102,7 @@ pub async fn schema_to_columns(
     let avro_schema = convert_avro(&json_schema, context).to_string();
     let schema = Schema::parse_str(&avro_schema).context("failed to parse avro schema")?;
     // TODO: do we need to support map type here?
-    avro_schema_to_column_descs(&schema, None)
+    avro_schema_to_column_descs(&schema, None).map_err(Into::into)
 }
 
 #[cfg(test)]
