+++ src/connector/src/parser/mod.rs
@@ -33,6 +33,7 @@ use risingwave_common::metrics::GLOBAL_ERROR_METRICS;
 use risingwave_common::types::{Datum, Scalar, ScalarImpl};
 use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_common::util::tracing::InstrumentStream;
+use risingwave_connector_encdec::decoder::avro::MapHandling;
 use risingwave_pb::catalog::{
     SchemaRegistryNameStrategy as PbSchemaRegistryNameStrategy, StreamSourceInfo,
 };
@@ -58,7 +59,6 @@ use crate::parser::util::{
     extract_header_inner_from_meta, extract_headers_from_meta, extreact_timestamp_from_meta,
 };
 use crate::schema::schema_registry::SchemaRegistryAuth;
-use crate::schema::InvalidOptionError;
 use crate::source::monitor::GLOBAL_SOURCE_METRICS;
 use crate::source::{
     extract_source_struct, BoxSourceStream, ChunkSourceStream, SourceColumnDesc, SourceColumnType,
@@ -1038,35 +1038,6 @@ pub struct AvroProperties {
     pub key_record_name: Option<String>,
     pub name_strategy: PbSchemaRegistryNameStrategy,
     pub map_handling: Option<MapHandling>,
-}
-
-/// How to convert the map type from the input encoding to RisingWave's datatype.
-///
-/// XXX: Should this be `avro.map.handling.mode`? Can it be shared between Avro and Protobuf?
-#[derive(Debug, Copy, Clone)]
-pub enum MapHandling {
-    Jsonb,
-    // TODO: <https://github.com/risingwavelabs/risingwave/issues/13387>
-    // Map
-}
-
-impl MapHandling {
-    pub const OPTION_KEY: &'static str = "map.handling.mode";
-
-    pub fn from_options(
-        options: &std::collections::HashMap<String, String>,
-    ) -> Result<Option<Self>, InvalidOptionError> {
-        let mode = match options.get(Self::OPTION_KEY).map(std::ops::Deref::deref) {
-            Some("jsonb") => Self::Jsonb,
-            Some(v) => {
-                return Err(InvalidOptionError {
-                    message: format!("unrecognized {} value {}", Self::OPTION_KEY, v),
-                })
-            }
-            None => return Ok(None),
-        };
-        Ok(Some(mode))
-    }
 }
 
 #[derive(Debug, Default, Clone)]
